+++ prkng/processing/common.py
@@ -61,7 +61,7 @@ CREATE TABLE slots
   way_name varchar,
   geom geometry(LineString,3857),
   geojson jsonb,
-  button_location hstore
+  button_location jsonb
 )
 """