+++ doc/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 WIP
 ---
 
+- use float type instead of string for button_location
 - use button_location coordinates for checkins
 - remove flask-login for the API